chore: remove unused blob function and clean up function signatures
- Deleted unused blob.ts function for todo storage
- Removed unused req parameter from scheduled-functions
- Updated cache purge endpoint from functions to API route
